Make Node 0.10 EOL official?
Hey folks! 

I'm not sure if you want to leave v0.10 in the table marked EOL for a while or if this should now be removed.

I'm raising this PR more as a place to clarify - v0.10 support ended yesterday but there wasn't any messaging on blog.nodejs.org. I had thought the discussion in the last meeting ended with an agreement that there was going to be more messaging. However, I do realise that it is time-consuming to do. 

Is there anything I can do to help?