Proposal to update staging branches
This is a proposal to update staging branches in a way that it should become easier to handle multiple release lines next to each other (semver-minor and patch releases). This is done by using two staging branches instead of one.

The minor one is the one in which each commit that should ever be backported should be targeted against. For patch releases we cut a release from the minor branch. That should a) reduce the conflicts for semver-minor releases and b) allow commits to land directly instead of having to wait until a release to be merged.

On top of that I loosened the rule that only LTS members should land PRs on the LTS branches. Having a few more people being able to do that seems like a good step forward. I personally would actually like to extend this by allowing any collaborator to land PRs against non-LTS staging branches.

I also changed the staging branches description not be generic. So far this only described the current state for the LTS staging branches even though the other release lines behaved AFAIK identical.

@nodejs/releasers PTAL