chore: move inactive releasers to emeritus
"Releasers have access to critical infrastructure in the project - this
elevated access must be restricted to active releasers. Members of the
releasers team should be offboarded when they no longer intend to prepare
releases. As a guideline, offboarding should be considered if a releaser has
not prepared a release in the past 12 months."

Refs: https://github.com/nodejs/Release/blob/main/GOVERNANCE.md#offboarding-releasers