update governance to reflect process
For some time all releasers have been allowed to make
LTS releases and backport changes. These documentation updates
are not intended to change any processes but rather to reflect
the current reality of how the team is operating.

There is still a backporters team that allows non-releasers to
backport commits, that is not currently documented.